validation user service

+++ src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -6,14 +10,25 @@ import { User } from '@prisma/client';
 import { PagnationParams } from 'src/utils/types/paginationParams';
 import { convertOrderby } from 'src/utils/functions/convertOrderby';
 import { ResponsePagination } from 'src/utils/types/responsePagination';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
+  saltOrRounds = 10;
   constructor(private readonly prisma: PrismaService) {}
 
   async create(data: CreateUserDto): Promise<User> {
+    const user = this.findByEmail(data.email);
+    if (user) {
+      throw new BadRequestException(
+        `User with email ${data.email} already exists`,
+      );
+    }
+    const newUser = { ...data };
+    const hash = await bcrypt.hash(data.password, this.saltOrRounds);
+    newUser.password = hash;
     return await this.prisma.user.create({
-      data,
+      data: newUser,
     });
   }
 
@@ -23,6 +38,11 @@ export class UserService {
     const order = convertOrderby(orderBy);
 
     const data = await this.prisma.user.findMany({
+      select: {
+        id: true,
+        email: true,
+        name: true,
+      },
       skip: (page - 1) * size,
       take: Number(size),
       orderBy: order,
@@ -45,6 +65,11 @@ export class UserService {
   }
 
   async update(id: string, data: UpdateUserDto): Promise<User> {
+    const user = await this.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
     return await this.prisma.user.update({
       data,
       where: { id: +id },
@@ -52,6 +77,11 @@ export class UserService {
   }
 
   async remove(id: string): Promise<User> {
+    const user = await this.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
     return await this.prisma.user.delete({
       where: { id: +id },
     });
